third_party/blink/web_tests/platform/generic/http/tests/cache/subresource-revalidation-referrer-expected.txt

Check that Referrer: for validation requests are correctly set.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS referrer.indexOf("subresource-revalidation-referrer.html") >= 0 is true
PASS referrer.indexOf("subresource-revalidation-referrer-iframe.html") >= 0 is true
PASS successfullyParsed is true

TEST COMPLETE
